Split Ghostty runtime helpers
